+++ nsarchive/utils/assets.py
@@ -2,7 +2,7 @@ import io
 import os
 from PIL import Image
 
-def open(path: str) -> Image:
+def open(path: str) -> bytes:
     curr_dir = os.path.dirname(os.path.abspath(os.path.join(__file__)))
     parent_dir = os.path.dirname(curr_dir)
     asset_path = os.path.join(parent_dir, 'assets', path)
