+++ requirments.txt
@@ -1,3 +1,3 @@
 # Modules nécessaires pour le fonctionnement du projet
-deta==1.2.0
+supabase
 pillow
